backend/requirements.txt
fastapi
openai-whisper
uvicorn
python-multipart
langchain
openai
python-dotenv
dotenv
ollama